+++ weblate/checks/models.py
@@ -110,6 +110,8 @@ class Check(models.Model):
     check = models.CharField(max_length=50, choices=CHECKS.get_choices())
     ignore = models.BooleanField(db_index=True, default=False)
 
+    weblate_unsafe_delete = True
+
     @cached_property
     def check_obj(self):
         try: