+++ weblate/trans/models/component.py
@@ -81,6 +81,7 @@ from weblate.trans.validators import (
 )
 from weblate.utils import messages
 from weblate.utils.celery import get_task_progress, is_task_ready
+from weblate.utils.db import FastDeleteMixin
 from weblate.utils.errors import report_error
 from weblate.utils.fields import JSONField
 from weblate.utils.licenses import get_license_choices, get_license_url, is_libre
@@ -161,7 +162,7 @@ class ComponentQuerySet(models.QuerySet):
         return self.exclude(repo__startswith="weblate:")
 
 
-class Component(models.Model, URLMixin, PathMixin):
+class Component(FastDeleteMixin, models.Model, URLMixin, PathMixin):
     name = models.CharField(
         verbose_name=gettext_lazy("Component name"),
         max_length=COMPONENT_NAME_LENGTH,